+++ src/main/java/kg/megalab/selim_trade/entity/enums/ERole.java
@@ -1,4 +1,4 @@
-package kg.megalab.selim_trade.entity;
+package kg.megalab.selim_trade.entity.enums;
 
 public enum ERole {
     ADMIN, SUPER_ADMIN
